fix items discussed
