+++ web/src/App.tsx
@@ -13,7 +13,7 @@ import AquaForms from './pages/AquaForms';
 import FormGenerator from './pages/FormGenerator';
 import AttestationAddresses from './pages/AttestationAddresses';
 import PdfSigner from './pages/PdfSigner';
-import WorkFlowPage from './pages/WorkFlow/WorkFlow';
+import WorkFlowPage from './pages/WorkFlow';
 // import Aquafier from 'aqua-js-sdk';
 declare global {
   interface Window {
@@ -46,7 +46,7 @@ function App() {
           <Route path="/loading" element={<Loading />} />
           <Route path="/share/:identifier" element={<SharePage />} />
           <Route path="/aqua-forms" element={<AquaForms /> } />
-          <Route path="/pdf-signer" element={<PdfSigner file={null}  submittingSignatureData={false} submitSignature={async(_one, _two)=>{
+          <Route path="/pdf-signer" element={<PdfSigner userCanSign={false} file={null}  submittingSignatureData={false} submitSignature={async(_one, _two)=>{
 
           }}/> } />
           <Route path="/workflow" element={<WorkFlowPage /> } />
